Add screen name to preconditions when the exception is throw for when the screen is not present in the backstack
