+++ magellan-library/src/main/java/com/wealthfront/magellan/Navigator.java
@@ -459,7 +459,7 @@ public class Navigator implements BackHandler {
     navigate(new HistoryRewriter() {
       @Override
       public void rewriteHistory(Deque<Screen> history) {
-        checkArgument(history.contains(screen), "Can't go back to a screen that isn't in history.");
+        checkArgument(history.contains(screen), String.format("Can't go back to a screen (%s) that isn't in history.", screen.getClass().getName()));
         while (history.size() > 1) {
           if (history.peek() == screen) {
             break;
