+++ magellan-library/src/test/java/com/wealthfront/magellan/DummyScreen.java
@@ -1,0 +1,18 @@
+package com.wealthfront.magellan;
+
+import android.content.Context;
+
+class DummyScreen extends Screen<BaseScreenView<DummyScreen>> {
+
+  private BaseScreenView<DummyScreen> view;
+
+  DummyScreen(BaseScreenView<DummyScreen> view) {
+    this.view = view;
+  }
+
+  @Override
+  protected BaseScreenView<DummyScreen> createView(Context context) {
+    return view;
+  }
+
+}
